pretty refactor
