+++ rust/candid/src/pretty/candid.rs
@@ -1,4 +1,4 @@
-use crate::pretty::*;
+use crate::pretty::utils::*;
 use crate::types::{Field, FuncMode, Function, Label, SharedLabel, Type, TypeEnv, TypeInner};
 use pretty::RcDoc;
 
@@ -219,7 +219,7 @@ pub fn compile(env: &TypeEnv, actor: &Option<Type>) -> String {
 #[cfg(feature = "value")]
 pub mod value {
     use super::{ident_string, pp_text};
-    use crate::pretty::*;
+    use crate::pretty::utils::*;
     use crate::types::value::{IDLArgs, IDLField, IDLValue};
     use crate::types::Label;
     use crate::utils::pp_num_str;
