+++ rust/candid/src/pretty/mod.rs
@@ -1,0 +1,4 @@
+//! pretty printer for Candid type and value
+
+pub mod candid;
+pub mod utils;
