+++ rust/candid/src/pretty/utils.rs
